Update README.md
SaveToken Making the difference

SaveToken.us is a project which wants to give back to the people. One of our key goals is to improve the standards of living of less privileged people all over the world. This will be done by donating 50% of all Savetoken’s transaction fees to 10 carefully selected projects that will make the world a better place, for example through fighting illnesses such as cancer.
Save Token will also create an intricate ecosystem with thousands of interconnected businesses who can rely on it both as a medium of exchange and as a store of value. Periodic burning of excess tokens will also ensure that its value will be consistently increasing. Because of the incredibly versatile platform Savetoken is creating, businesses will be able to easily manage key information such as research information as well as allowing easy access to key statistics like revenues.
Our aim is for Save Token to become the principal cryptocurrency which will largely replace fiat currencies over the course of several years. Save Token will serve as a solution to a prominent problem, becoming a highly used and valuable token for millions of businesses and individuals. The implementation of Zcash’s zk-snarks privacy technology will also mean optional privacy for those firms that desire it. This whitepaper will explain the details of this vision and exactly how we see it coming to fruition.
Smart contract is published in Ethereum blockchain.
Address is 0xc1eEcf1f4AF8EB9a2a19f6C26B434aA96ce859e1